Removed Dependency on LinkList
- removed dependency of LinkList and made RightExpression a linked list
node.  This should make it easier to create a stack-based evaluator
because we can just push a RightExpression on the stack and evaluated
it from left to right as we go.  This would have been harder before
because LinkList was not an Expression, so we could not push it onto an
expression stack.